requirements.txt
#python 3.13.7
fastapi
uvicorn[standard]
python-dotenv
pydantic
google-generativeai